Remove some reduntant cast warning in ProtectViewWorker/RequestHandler (OFBIZ-4102)
git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1081064 13f79535-47bb-0310-9956-ffa450edef68

+++ framework/webapp/src/org/ofbiz/webapp/control/ProtectViewWorker.java
@@ -94,7 +94,7 @@ public class ProtectViewWorker {
                         durationByViewAccessed.put(viewNameUserLoginId, now + maxHitsDuration);
                     } else {
                         Long maxDuration = durationByViewAccessed.get(viewNameUserLoginId);
-                        Long newMaxHits = (Long) curMaxHits + one;
+                        Long newMaxHits = curMaxHits + one;
                         hitsByViewAccessed.put(viewNameUserLoginId, newMaxHits);
                         // Are we in a period of time where we need to check if there was too much hits ?
                         if (now < maxDuration) {
